Implement support for grabbing a users groups
This will walk the group list for a user and attempt to grab all of the
group **names** that a user is in, as we're far more interested in the
names than the IDs.

By default we'll attempt to grab 5 groups, however if there are more then
we'll reallocate to the number of groups the user is in, grab the list again
and try to populate the user struct with all the IDs.

Signed-off-by: Ikey Doherty <ikey@solus-project.com>